Refactored backend code
