+++ backend/middlewares/requestLogger.js
@@ -1,8 +1,6 @@
 const logger = (req, res, next) => {
     const start = Date.now()
-    const logRequest = `[${new Date().toISOString()}] ${req.method} ${
-        req.originalUrl
-    }`
+    const logRequest = `[${new Date().toISOString()}] ${req.method} ${req.originalUrl}`
 
     res.on("finish", () => {
         const duration = Date.now() - start
