+++ frontend/src/pages/Signup.jsx
@@ -49,7 +49,7 @@ function Signup() {
 
     const handleGoogleSignup = () => {
         axiosInstance
-            .post("/auth/google/request")
+            .post("/auth/google/redirect")
             .then((res) => (window.location.href = res.data.url))
             .catch((error) => {
                 console.error("Error checking authentication status:", error)
