+++ frontend/src/signals/user.js
@@ -3,12 +3,6 @@ import { effect, signal } from "@preact/signals-react"
 import { getCookie } from "@/helpers/cookies"
 
 const isUserLoggedIn = () => {
-    // const accessToken = getCookie("accessToken")
-    // const refreshToken = getCookie("refreshToken")
-
-    // if (accessToken === null && refreshToken === null) return false
-    // if (accessToken === "null" && refreshToken === "null") return false
-
     const refreshTokenId = getCookie("refreshTokenId")
     if (refreshTokenId === null) return false
     if (refreshTokenId === "null") return false
